re-styled the graph
