+++ package.json
@@ -87,6 +87,7 @@
     "@storybook/addons": "^5.2.6",
     "@storybook/react": "^5.2.6",
     "@types/react-router-dom": "^5.1.7",
+    "@types/vis": "^4.21.21",
     "babel-eslint": "^10.0.3",
     "dotenv": "^8.2.0",
     "eslint-config-airbnb": "^18.2.1",
