+++ src/styles/colors.js
@@ -11,6 +11,7 @@ export const COLOR = {
   blue: `#307fe6`,
   blueDark: `#0065E5`,
   blueLight: `#45a9ec`,
+  blueExtraLight: `#a3d4f5`,
   cyan: `#98dbef`,
   darkGrey: `#807c7c`,
   green: `#63C13C`,
@@ -23,13 +24,16 @@ export const COLOR = {
   pink: `#e543b4`,
   pinkLight: `#FF5CA2`,
   purple: `#C657D0`,
+  darkPurple: '#4b1650',
   red: `#e74c3c`,
   redPale: `#FF8974`,
   transparentBlack: `#00000073`,
+  transparentWhite: `#ffffffdd`,
   transparentGrey: `#bfbfbf21`,
   turquoise: `#36DFB9`,
   white: `white`,
   yellow: `#eeda13`,
+  yellowPale: '#f6ed88',
   darkCharcoal: `#333333`,
 };
 
@@ -121,13 +125,13 @@ export const COLOR_LAYOUT = {
 export const COLOR_TASK_STATUS = {
   [TASK.ACTIVE]: COLOR.blueLight,
   [TASK.COMPLETED]: COLOR.greenLight,
-  [TASK.CRASHED]: COLOR.pink,
-  [TASK.CREATING]: COLOR.yellow,
+  [TASK.CRASHED]: COLOR.darkCharcoal,
+  [TASK.CREATING]: COLOR.grey,
   [TASK.FAILED]: COLOR.red,
-  [TASK.PENDING]: COLOR.darkGrey,
+  [TASK.PENDING]: COLOR.lightGrey,
   [TASK.PRESCHEDULE]: COLOR.cyan,
   [TASK.SKIPPED]: COLOR.orangeLight,
-  [TASK.STALLED]: COLOR.grey,
+  [TASK.STALLED]: COLOR.darkGrey,
   [TASK.SUCCEED]: COLOR.greenLight,
   [TASK.WARNING]: COLOR.orange,
 };
